feat: add qsub_job_monitor and combine_categorical_runs to channel
this was a test run to determine whether it was feasible to add new packages collaboratively with a
standard clone/commit/push process; and it seems like the answer is yes! this adds qsub_job_monitor
and combine_categorical_runs to the channel, and messes with some structural things to make the
commit cleaner.