feat: add ldsc package because the public one is trash
for uncharacterized reasons, the public ldsc conda package has a locked version for a dependency
that no longer exists in conda, so the installation fails. this ldsc package is literally just the
same thing as the public one, literally with their exact same recipe, except I removed the version
restriction. for reference, I have had a local ldsc conda environment for ages, based on the
environment released with ldsc itself, and the dependency was happily not the locked version and
worked fine the whole time. so you know, one of those mysteries that crop up in bioinformatics